♻️ refactor: 경력/학력 입력/수정 페이지 로직 추상화 #429
- 경력/학력 입력/수정 페이지 내 상태 관리, API 호출, 유효성 검사, 이벤트 핸들러 등을 커스텀 훅으로 캡슐화
- 페이지 파일은 UI를 렌더링하는데 집중
